Updated structure to remove folder nesting
