+++ src/Charts/DifferenceChart.tsx
@@ -1,9 +1,9 @@
 import { h, Component } from 'preact';
-import { Margin } from '../../types';
-import { Axis } from '../../Axis';
+import { Margin } from '../types';
+import { Axis } from '../Axis';
 import { scaleLinear, scaleBand, ScaleBand, ScaleLinear } from 'd3-scale';
 import { max } from 'd3-array';
-import { pluck } from '../../Utils/pluck';
+import { pluck } from '../Utils/pluck';
 
 declare const ResizeObserver: any;
