+++ src/Charts/HorizontalBar.tsx
@@ -1,53 +1,54 @@
 import { h, Component } from 'preact';
-import { Margin, GroupedDataObject } from '../../types';
-import { Axis } from '../../Axis';
+import { Margin, GroupedDataObject } from '../types';
+import { Axis } from '../Axis';
 import { scaleLinear, scaleBand, ScaleBand, ScaleLinear, scaleOrdinal, ScaleOrdinal } from 'd3-scale';
 import { max } from 'd3-array';
-import { pluckUnique } from '../../Utils/pluck';
-import { colourArray } from '../../colors';
+import { pluckUnique } from '../Utils/pluck';
+import { colourArray } from '../colors';
 
 declare const ResizeObserver: any;
 
-interface GroupedBarProps {
+interface HorizontalBarProps {
     name: string;
     data: GroupedDataObject;
     groups: string[];
     legendReference: {[key: string]: string};
+
     height?: number;
     width?: number;
     margin?: Margin;
     ticks?: number;
 }
 
-interface GroupedBarDefaultProps {
+interface HorizontalBarDefaultProps {
     height?: number;
     width?: number;
     margin?: Margin;
     ticks?: number;
 }
 
-interface GroupedBarState {
+interface HorizontalBarState {
     width: number;
     innerWidth: number;
     height: number;
     innerHeight: number;
 }
-export class GroupedBar extends Component<GroupedBarProps, GroupedBarState> {
-    public static defaultProps: GroupedBarDefaultProps = {
-        height: 500,
-        width: 500,
+export class HorizontalBar extends Component<HorizontalBarProps, HorizontalBarState> {
+    public static defaultProps: HorizontalBarDefaultProps = {
+        height: 800,
+        width: 600,
         margin: {
             top: 25,
             right: 25,
-            bottom: 75,
-            left: 50,
+            bottom: 50,
+            left: 150,
         },
         ticks: 6,
     };
     private chartSVG: HTMLBaseElement;
     private resizeOb: any;
 
-    constructor (props: GroupedBarProps) {
+    constructor (props: HorizontalBarProps) {
         super(props);
         const innerWidth = props.width - props.margin.right - props.margin.left;
         const innerHeight = props.height - props.margin.top - props.margin.bottom;
@@ -58,50 +59,51 @@ export class GroupedBar extends Component<GroupedBarProps, GroupedBarState> {
             innerWidth,
         };
     }
-    public render ({ margin, ticks, data, groups, legendReference, name }: GroupedBarProps,
-                   { height, width, innerHeight, innerWidth }: GroupedBarState) {
-        let yMax = 0;
+    public render ({ margin, ticks, data, groups, name, legendReference }: HorizontalBarProps,
+                   { height, width, innerHeight, innerWidth }: HorizontalBarState) {
+        let xMax = 0;
 
         for (const key of groups) {
             const groupMax = max(data[key], (d) => d.value);
-            yMax = groupMax > yMax ? groupMax : yMax;
+            xMax = groupMax > xMax ? groupMax : xMax;
         }
 
         const names = pluckUnique(data[groups[0]], 'name') as string[];
 
-        const yScale = scaleLinear()
-            .range([innerHeight, 0])
-            .domain([0, yMax])
+        const xScale = scaleLinear()
+            .range([0, innerWidth])
+            .domain([0, xMax])
             .nice();
 
-        const xScale = scaleBand()
-            .rangeRound([0, innerWidth])
+        const yScale = scaleBand()
+            .rangeRound([innerHeight, 0])
             .paddingInner(0.1)
             .domain(groups);
 
-        const x1 = scaleBand()
+        const y1 = scaleBand()
             .padding(0.05)
             .domain(names)
-            .rangeRound([0, xScale.bandwidth()]);
+            .rangeRound([0, yScale.bandwidth()]);
 
         const colourScale = scaleOrdinal(colourArray);
         return (
             <svg ref={(svg) => this.chartSVG = svg} class={name} height={height} width={width}>
                 <g transform={`translate(${margin.left}, ${margin.top})`}>
-                    <Axis height={innerHeight} axisType='x' scale={xScale} rotateScaleText={true}/>
-                    <Axis width={innerWidth} axisType='y' scale={yScale} grid={true} ticks={ticks}/>
+                    <Axis height={innerHeight} axisType='x' scale={xScale} rotateScaleText={false} grid={true}/>
+                    <Axis width={innerWidth} axisType='y' scale={yScale} ticks={ticks}/>
                     {
                         data &&
-                            this.createBars(xScale, x1, yScale, innerHeight, groups, colourScale)
+                            this.createBars(yScale, y1, xScale, groups, colourScale)
                     }
                     {
                         names &&
                             names.map((barName, idx) =>
                                 <g transform={`translate(0, ${idx * 20})`}>
-                                    <rect x={innerWidth + margin.right - 18} width={18} height={15}
+                                    <rect x={innerWidth + margin.right - 18} width={18} height={15} stroke='black'
                                         strokeWidth='1px' fill={colourScale(barName)}>
                                     </rect>
-                                    <text x={innerWidth + margin.right - 24} y={9} dy='0.35em' text-anchor='end'>
+                                    <text x={innerWidth + margin.right - 24} y={9} dy='0.35em' fill='whitesmoke'
+                                        text-anchor='end'>
                                         {legendReference[barName]}
                                     </text>
                                 </g>)
@@ -139,14 +141,14 @@ export class GroupedBar extends Component<GroupedBarProps, GroupedBarState> {
         this.setState({innerWidth, innerHeight, height, width});
     }
 
-    private createBars = (x0: ScaleBand<string>, x1: ScaleBand<string>, y: ScaleLinear<number, number>,
-                          height: number, groups: string[], colourScale: ScaleOrdinal<string, string>) => {
+    private createBars = (y0: ScaleBand<string>, y1: ScaleBand<string>, x: ScaleLinear<number, number>,
+                          groups: string[], colourScale: ScaleOrdinal<string, string>) => {
         return groups.map((group) =>
-            <g transform={`translate(${x0(group)}, 0)`}>
+            <g transform={`translate(0, ${y0(group)})`}>
                 {
                     this.props.data[group].map((entry) =>
-                        <rect width={x1.bandwidth()} x={x1(entry.name as string)} y={y(entry.value)}
-                            height={height - y(entry.value)} fill={colourScale(entry.name as string)}
+                        <rect height={y1.bandwidth()} x={0} y={y1(entry.name as string)}
+                            width={x(entry.value)} fill={colourScale(entry.name as string)}
                             title={entry.value.toFixed(4)}>
                         </rect>)
                 }
