+++ src/Charts/LineScatter.tsx
@@ -1,12 +1,12 @@
 import { h, Component } from 'preact';
-import { Axis } from '../../Axis';
-import { Margin, DataArray, NumberObject } from '../../types';
+import { Axis } from '../Axis';
+import { Margin, DataArray, NumberObject } from '../types';
 import { scaleLinear } from 'd3-scale';
 import { extent } from 'd3-array';
 import { select, event } from 'd3-selection';
 import { brush } from 'd3-brush';
 import { line, curveNatural } from 'd3-shape';
-import { colourArray } from '../../colors';
+import { colourArray } from '../colors';
 
 declare const ResizeObserver: any;
