+++ src/preactCharts.ts
@@ -6,7 +6,7 @@ import { LineScatter } from './Charts/LineScatter';
 import { RangeChart } from './Charts/RangeChart';
 import { ScatterPlot } from './Charts/ScatterPlot';
 import { TrendChart } from './Charts/TrendChart';
-import { Flag } from './Charts/TrendChart/Flag';
+import { Flag } from './Children/Flag';
 export * from './types';
 export {
     DifferenceChart,
